refactor generically typed checkpoints
